Image

+++ spec/requests/api/v1/weather_request_spec.rb
@@ -33,7 +33,7 @@ describe "Weather API Endpoints" do
       expect(current_weather).to have_key(:temperature)
       expect(current_weather[:temperature]).to be_a(Float)
       expect(current_weather).to have_key(:feels_like)
-      expect(current_weather[:feels_like]).to be_a(Float)
+      expect(current_weather[:feels_like]).to be_a(Numeric)
       expect(current_weather).to have_key(:humidity)
       expect(current_weather[:humidity]).to be_a(Numeric)
       expect(current_weather).to have_key(:uvi)
